# docs/architecture/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "A股/港股/美股自選股智慧分析系統 API\n\n## 功能模組\n- 股票分析：觸發 AI 智慧分析\n- 歷史記錄：查詢歷史分析報告\n- 股票資料：獲取行情資料\n\n## 認證方式\n當前版本暫無認證要求",
    "version": "1.0.0",
    "contact": {
      "name": "Daily Stock Analysis Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地開發伺服器"
    }
  ],
...
  "paths": {
...
    "/api/v1/history": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "獲取歷史分析列表",
        "description": "分頁獲取歷史分析記錄摘要，支援按股票程式碼和日期範圍篩選",
        "operationId": "getHistoryList",
        "parameters": [
          {
            "name": "stock_code",
            "in": "query",
            "description": "股票程式碼篩選",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "開始日期 (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "結束日期 (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "頁碼（從 1 開始）",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每頁數量",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "歷史記錄列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{query_id}": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "獲取歷史報告詳情",
        "description": "根據 query_id 獲取完整的歷史分析報告",
        "operationId": "getHistoryDetail",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "description": "分析記錄唯一標識",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "報告詳情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "404": {
            "description": "報告不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{query_id}/news": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "獲取歷史報告關聯新聞",
        "description": "根據 query_id 獲取關聯的新聞情報列表（為空也返回 200）",
        "operationId": "getHistoryNews",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "description": "分析記錄唯一標識",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返回數量限制",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "新聞情報列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsIntelResponse"
                }
              }
            }
          },
          "500": {
            "description": "伺服器錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HistoryListResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "總記錄數"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            }
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "items"
        ]
      },
      "NewsIntelItem": {
        "type": "object",
        "description": "新聞情報條目",
        "properties": {
          "title": {
            "type": "string",
            "description": "新聞標題"
          },
          "snippet": {
            "type": "string",
            "description": "新聞摘要（最多50字）"
          },
          "url": {
            "type": "string",
            "description": "新聞連結"
          }
        },
        "required": [
          "title",
          "url"
        ]
      },
      "NewsIntelResponse": {
        "type": "object",
        "description": "新聞情報響應",
        "properties": {
          "total": {
            "type": "integer",
            "description": "新聞條數"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsIntelItem"
            },
            "description": "新聞列表"
          }
        },
        "required": [
          "total",
          "items"
        ]
      },
      "HistoryItem": {
        "type": "object",
        "description": "歷史記錄摘要（列表展示用）",
        "properties": {
          "query_id": {
            "type": "string"
          },
          "stock_code": {
            "type": "string"
          },
          "stock_name": {
            "type": "string"
          },
          "report_type": {
            "type": "string"
          },
          "sentiment_score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "operation_advice": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "query_id",
          "stock_code",
          "created_at"
        ]
      },
      "AnalysisReport": {
        "type": "object",
        "description": "完整分析報告",
        "properties": {
          "meta": {
            "type": "object",
            "description": "元資訊",
            "properties": {
              "query_id": {
                "type": "string"
              },
              "stock_code": {
                "type": "string"
              },
              "stock_name": {
                "type": "string"
              },
              "report_type": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "model_used": {
                "type": "string",
                "description": "分析使用的 LLM 模型（完整名，如 gemini/gemini-2.0-flash）"
              }
            }
          },
          "summary": {
            "type": "object",
            "description": "概覽區（首屏展示）",
            "properties": {
              "analysis_summary": {
                "type": "string",
                "description": "關鍵結論"
              },
              "operation_advice": {
                "type": "string",
                "description": "操作建議"
              },
              "trend_prediction": {
                "type": "string",
                "description": "趨勢預測"
              },
              "sentiment_score": {
                "type": "integer",
                "description": "情緒評分 (0-100)"
              },
              "sentiment_label": {
                "type": "string",
                "description": "情緒標籤",
                "enum": [
                  "極度悲觀",
                  "悲觀",
                  "中性",
                  "樂觀",
                  "極度樂觀"
                ]
              }
            }
          },
          "strategy": {
            "type": "object",
            "description": "策略點位區",
            "properties": {
              "ideal_buy": {
                "type": "string",
                "description": "理想買入價"
              },
              "secondary_buy": {
                "type": "string",
                "description": "第二買入價"
              },
              "stop_loss": {
                "type": "string",
                "description": "止損價"
              },
              "take_profit": {
                "type": "string",
                "description": "止盈價"
              }
            }
          },
          "details": {
            "type": "object",
            "description": "詳情區（可摺疊）",
            "properties": {
              "news_content": {
                "type": "string",
                "description": "新聞摘要"
              },
              "raw_result": {
                "type": "object",
                "description": "原始分析結果（JSON）"
              },
              "context_snapshot": {
                "type": "object",
                "description": "分析時上下文快照（JSON）"
              }
            }
          }
        },
        "required": [
          "meta",
          "summary"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "錯誤型別"
          },
          "message": {
            "type": "string",
            "description": "錯誤詳情"
          },
          "detail": {
            "type": "object",
            "description": "附加錯誤資訊"
          }
        },
        "required": [
          "error",
          "message"
        ]
      }
    }
  }
}